feat(Node): add the Node class implementation
This patch adds the Node Class Implementation. It's a simple class
that just have a constructor that sets 2 properties:

- `value`: which can be any meaningful primitive data.
- `next`: which is a pointer to the next Node reference.

This class is used to pass nodes to add to the SinglyLinkedList
instances. Ideally that should be internal and hidden from the
consumer but due to time constraints this is how it is for now.